Add experimental support for repeating the replacement also in visual mode through visualrepeat.vim
Renamed vmap <Plug>ReplaceWithRegisterOperator to <Plug>ReplaceWithRegisterVisual for that. A repeat in visual mode will now apply the previous line and operator replacement to the selection text. A repeat in normal mode will apply the previous visual mode replacement at the current cursor position, using the size of the last visual selection.